requirements.txt
git+https://github.com/pytorch/tnt.git@master
fire
numpy
torchvision
visdom
pytorch
future
ipdb